Update InstallCache.rst
Question: The last step for configuring Cache is to enable TELNET, but a previous step said that it wasn't necessary to use use a licensed version for testing. What are the ramifications of not having TELNET services? Should the instructions tell the user to skip on down to installing CACHE.DAT if they are not using a licensed version of CACHE at this point?

Standardized use of "macOS" to conform with preferred Apple usage.